@@ -3,13 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "5.12.0",
-    "@testing-library/react": "11.2.6",
-    "@testing-library/user-event": "13.1.8",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "^4.0.3",
-    "web-vitals": "^1.1.1"
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/user-event": "14.4.3",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
