@@ -3,13 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "5.11.4",
-    "@testing-library/react": "11.0.4",
-    "@testing-library/user-event": "12.1.6",
-    "axios": "^0.21.0",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-scripts": "^4.0.0"
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/user-event": "14.4.3",
+    "axios": "^1.3.4",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
