@@ -3,14 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "5.11.4",
-    "@testing-library/react": "11.0.4",
-    "@testing-library/user-event": "12.1.6",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-player": "2.6.2",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "^4.0.0"
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/user-event": "14.4.3",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-player": "2.11.2",
+    "react-router-dom": "6.8.1",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
