@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Switch, Route } from "react-router-dom";
+import { Routes, Route } from "react-router-dom";
 import "./App.css";
 
 import Navbar from "./components/Navbar/Navbar";
@@ -12,10 +12,10 @@ import NotFound from "./components/NotFound/NotFound";
 
 /* 
   1. Install "react-router-dom". (In your terminal)
-  2. <BrowserRouter> --> Helps rendering components passed to it. (wrapped around <App> in index.js)
+  2. <BrowserRouter> --> Stores the current browser location and navigates . (wrapped around <App> in index.js)
   3. Create components to render. (Components folder)
-  4. <Switch> --> Ensures that only one component is rendered at a time.
-  5. <Route> --> Links between the components. Placed between <Switch> element.
+  4. <Routes> --> Matches the current location with one of the child routes.
+  5. <Route> --> Couples URL location with React component. Placed between <Routes> element.
   6. <Link> --> Attach clickable links to different pages. (In Navbar component)
   7. <NavLink> --> special version of <Link> that adds styling attributes. (In Navbar component)
 */
@@ -24,18 +24,18 @@ const App = () => {
   return (
     <section className="container is-fluid">
       <Navbar />
-      <Switch>
-        <Route exact path="/" component={Home} />
-        <Route path="/about" component={About} />
-        <Route exact path="/projects" component={Projects} />
-        <Route path="/projects/:id" component={ProjectDetails} />
-        <Route component={NotFound} />
-      </Switch>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/about" element={<About />} />
+        <Route path="/projects" element={<Projects />} />
+        <Route path="/projects/:id" element={<ProjectDetails />} />
+        <Route path="*" element={<NotFound />} />
+      </Routes>
 
-      {/* <Switch>
-        <Route path="/about" component={About} />
-        <Route path="/" component={Home} />
-      </Switch> */}
+      {/* <Routes>
+        <Route path="/about" element={About} />
+        <Route path="/" element={Home} />
+      </Routes> */}
     </section>
   );
 };
